Add getRitualInitiationCost to CoordinatorAgent and update acceptance tests accordingly; this way we don't care about fee values.
